Enable -Wfloat-conversion and fix warnings (#2734)
* Enable -Wfloat-conversion and fix warnings

* replace round with the right rint (faster?)

* indeed it's just u8

* merge other Niilo's PR

+++ src/lib/fixed_fft_r2.c
@@ -207,7 +207,7 @@ static void InitW(intFFTr2_t *pIntFFT) {
     tW = &(pIntFFT->W[_bsize]);
 
     for (k = 0; k < _bsize; k++) {
-      phase = (-2.0 * M_PI * k * nBlocks) / (pIntFFT->N);
+      phase = (-2.0f * (float)M_PI * k * nBlocks) / (pIntFFT->N);
       c = floorf(fScale * cosf(phase));
       s = floorf(fScale * sinf(phase));
