+++ src/utils/signal_db/signal_db.c
@@ -392,7 +392,7 @@ float code_to_tcxo_doppler_min(code_t code) {
   float doppler;
 
   /* use worst case Doppler */
-  doppler = -TCXO_FREQ_OFFSET_MAX_PPM * GLO_L1_TCXO_PPM_TO_HZ;
+  doppler = (float)(-TCXO_FREQ_OFFSET_MAX_PPM * GLO_L1_TCXO_PPM_TO_HZ);
 
   return doppler;
 }
@@ -408,7 +408,7 @@ float code_to_tcxo_doppler_max(code_t code) {
   float doppler;
 
   /* use worst case Doppler */
-  doppler = TCXO_FREQ_OFFSET_MAX_PPM * GLO_L1_TCXO_PPM_TO_HZ;
+  doppler = (float)(TCXO_FREQ_OFFSET_MAX_PPM * GLO_L1_TCXO_PPM_TO_HZ);
 
   return doppler;
 }
